bug fix and add title

+++ src/components/Artc.vue
@@ -38,14 +38,15 @@
 		},
 		methods: {
 			updArtc(value) {
-				if (value)
-					axios({url: `http://localhost:3000/api/articles/${this.post.slug}/favorite`, method: 'delete', headers: authHeader()})
-						.then(response => {this.post = response.data.article})
-						.catch(e => console.log(e))
-				else
-					axios({url: `http://localhost:3000/api/articles/${this.post.slug}/favorite`, method: 'post', headers: authHeader()})
-						.then(response => {this.post = response.data.article})
-						.catch(e => console.log(JSON.stringify(e)))
+				if (authHeader())
+					if (value)
+						axios({url: `http://localhost:3000/api/articles/${this.post.slug}/favorite`, method: 'delete', headers: authHeader()})
+							.then(response => {this.post = response.data.article})
+							.catch(e => console.log(e))
+					else
+						axios({url: `http://localhost:3000/api/articles/${this.post.slug}/favorite`, method: 'post', headers: authHeader()})
+							.then(response => {this.post = response.data.article})
+							.catch(e => console.log(JSON.stringify(e)))
 			},
 			returnTag(value) {
 				this.$emit('changeTag', value);
